requirements.txt
aiogram==3.4.1
supabase==2.3.1
python-dotenv==1.0.0
asyncio==3.4.3
pandas==2.1.4
asyncpg==0.29.0
loguru==0.7.2
aiofiles==23.2.1
python-dateutil==2.8.2 